Final commit Added css styles and changed the html files.
I added the styles for the print css style sheet. Fixed the links for the html pages so the links on the web page will work correctly. Added the comments for the css and the html files. Added the images width for the print css stylesheet. The style is for the section with the id of sheet and the rule is to make the width 6inches and has margin of 0in for top/bottom and auto for left/right.The style rule is for the nav with the class of horizontal and for the navigation, aside, and footer content is not to show up on the printed version of the web page.  Added the tablet and Desktop screen styles. ready for review.